integration_tests/requirements.txt
robotframework
docker
robotframework-pythonlibcore
jinja2